unit of work and local generic repository commit method changed

+++ DesignPatternsAndPrinciples/DesignPatternsAndPrinciplesConsole/Program.cs
@@ -17,12 +17,14 @@ public class Program
             })
             .AddScoped(typeof(IRepository<>), typeof(LocalRepository<>))
             .AddScoped(typeof(IUnitOfWorkSession), typeof(UnitOfWorkSession))
+            .AddScoped(typeof(IContextHandler), typeof(ContextHandler))
             .BuildServiceProvider();
 
         Console.WriteLine("Hello world..");
 
         using (var session = serviceProvider.GetService<IUnitOfWorkSession>())
         using (var repository = session.GetRepository<Foo>())
+        using (var barRepository = session.GetRepository<Bar>())
         {
             var data = repository.Get(Guid.NewGuid());
 
@@ -31,7 +33,7 @@ public class Program
             var newData = repository.Get(jackId);
             if (newData == null)
             {
-                session.Commit();
+                //session.Commit();
                 newData = repository.Get(jackId);
                 if (newData != null)
                 {
@@ -39,12 +41,15 @@ public class Program
                 }
             }
 
-            repository.Delete(newData);
+            if (newData != null)
+            {
+                repository.Delete(newData);
+            }
 
             var deletedData = repository.Get(jackId);
             if (deletedData != null)
             {
-                session.Commit();
+                //session.Commit();
                 deletedData = repository.Get(jackId);
                 if (deletedData == null)
                 {
@@ -54,10 +59,15 @@ public class Program
 
             var elonId = Guid.NewGuid();
             repository.Insert(new Foo() { Id = elonId, Name = "Elon" });
-            session.Commit();
+            //session.Commit();
 
             var list = repository.FindAll(w => w.Id == elonId);
+
+            barRepository.Insert(new Bar { Id = Guid.NewGuid(), Content = "Lorem Ipsum is simply dummy text of the printing and typesetting industry." });
+            //session.Commit();
         }
+
+        // Eğer hiç commit yapılma ise session biterken(dispose edilirken) commit olur.
 
         Console.WriteLine("Bye world..");
         Console.ReadKey();
@@ -69,5 +79,12 @@ public class Program
         public Guid Id { get; set; }
 
         public string Name { get; set; }
+    }
+
+    public class Bar : IEntity
+    {
+        public Guid Id { get; set; }
+
+        public string Content { get; set; }
     }
 }
